requirements.txt
qrcode
gradio
Pillow
